requirements.txt
djangorestframework
psycopg2-binary
python-decouple
djangorestframework-simplejwt
pillow
drf-yasg
django-cors-headers